add: gtest debugging
fix: cuda forward linear layer test. I was trying to allocate data
directly on a CUDA storage. Now I allocate on CPU and then move on GPU.
It's slow but I don't care it's just for the test units
todo: make something like torch.randn() which is helpful in general